control/controlclient: treat status 4XX errors as user visible
For auth keys with tags, it is possible that use of the key
results in a 400 error because the tag associated with the key
no longer exists (it existed when the key was originally created).

Prior to this change, `tailscale up -authkey=...` would loop
indefinitely without any problems reported to the user.

Signed-off-by: Joe Tsai <joetsai@digital-static.net>